feat: cria novo módulo profile com rotas de profile
