+++ internal/services/client/routes.go
@@ -11,6 +11,7 @@ import (
 	"github.com/caio-bernardo/dragonite/internal/notifier"
 	"github.com/caio-bernardo/dragonite/internal/repository"
 	"github.com/caio-bernardo/dragonite/internal/services/client/auth"
+	"github.com/caio-bernardo/dragonite/internal/services/client/profile"
 	"github.com/caio-bernardo/dragonite/internal/services/client/rooms"
 	"github.com/caio-bernardo/dragonite/internal/types"
 	"github.com/caio-bernardo/dragonite/internal/util"
@@ -34,6 +35,7 @@ func (h *Handler) RegisterRoutes(mux *http.ServeMux, authMiddleware types.Middle
 
 	auth := auth.NewHandler(h.userStore, h.deviceStore)
 	roomHandler := rooms.NewHandler(h.canalStore, h.usuarioCanalStore, h.eventoStore, os.Getenv("SERVER_NAME"), h.notifier)
+	profileHandler := profile.NewHandler(h.userStore)
 
 	mux.HandleFunc("GET /_matrix/client/versions", h.getVersions)
 
@@ -42,24 +44,12 @@ func (h *Handler) RegisterRoutes(mux *http.ServeMux, authMiddleware types.Middle
 
 	// sincronização de dados
 	mux.Handle("GET /_matrix/v3/client/sync", authMiddleware(http.HandlerFunc(h.syncClient))) // WARN: esse é o dificil
-
-	// chats e manipulação de salas
-	roomHandler.RegisterRoutes(mux, authMiddleware)
-
 	// busca de usuários
 	mux.Handle("POST /_matrix/client/v3/user_directory/search", authMiddleware(http.HandlerFunc(h.searchUsers)))
 
-	// Perfil do usuário
-	mux.HandleFunc("GET /_matrix/client/v3/profile/{userId}", h.getProfile)
-
-	// Chave do perfil do usuário
-	mux.HandleFunc("GET /_matrix/client/v3/profile/{userId}/keys", h.getProfileKey)
-
-	// Alterar chave do perfil do usuário
-	mux.HandleFunc("PUT /_matrix/client/v3/profile/{userId}/keys", h.putProfileKey)
-
-	// Remover chave do perfil do usuário
-	mux.HandleFunc("DELETE /_matrix/client/v3/profile/{userId}/keys", h.deleteProfileKey)
+	// chats e manipulação de salas
+	roomHandler.RegisterRoutes(mux, authMiddleware)
+	profileHandler.RegisterRoutes(mux, authMiddleware)
 }
 
 func (h *Handler) getVersions(w http.ResponseWriter, r *http.Request) {
@@ -187,168 +177,4 @@ func (h *Handler) syncClient(w http.ResponseWriter, r *http.Request) {
 	response := encodeEventsIntoResponse(events, newToken)
 
 	util.WriteJSON(w, http.StatusOK, response)
-}
-
-// mapMatrixKeyToDB converte a chave do Matrix para a coluna do bd.
-// Isso evita SQL Injection, pois não usamos a string do usuário direto na query.
-func mapMatrixKeyToDB(keyName string) string {
-	switch keyName {
-	case "displayname":
-		return "nome_usuario"
-	case "avatar_url":
-		return "foto_usuario"
-	default:
-		return ""
-	}
-}
-
-// GET /_matrix/client/v3/profile/{userId}
-func (h *Handler) getProfile(w http.ResponseWriter, r *http.Request) {
-	userID := r.PathValue("userId")
-
-	usuario, err := h.userStore.GetNameAndPhotoByID(r.Context(), userID)
-
-	if err != nil {
-		if err == types.ErrNotFound {
-			util.WriteError(w, http.StatusNotFound, types.ErrorResponse{
-				ErrCode: "M_NOT_FOUND",
-				Message: "Profile not found",
-			})
-			return
-		}
-		util.WriteError(w, http.StatusInternalServerError, types.ErrorResponse{
-			ErrCode: "M_UNKNOWN",
-			Message: "Internal server error",
-		})
-		return
-	}
-
-	response := model.ProfileResponse{
-		DisplayName: usuario.Nome,
-		AvatarURL:   usuario.Foto,
-	}
-
-	util.WriteJSON(w, http.StatusOK, response)
-}
-
-// GET /_matrix/client/v3/profile/{userId}/{keyName}
-func (h *Handler) getProfileKey(w http.ResponseWriter, r *http.Request) {
-	userID := r.PathValue("userId")
-	keyName := r.PathValue("keyName") // Vai ser "displayname" ou "avatar_url"
-
-	usuario, err := h.userStore.GetNameAndPhotoByID(r.Context(), userID)
-
-	if err != nil {
-		if err == types.ErrNotFound {
-			util.WriteError(w, http.StatusNotFound, types.ErrorResponse{
-				ErrCode: "M_NOT_FOUND",
-				Message: "User not found",
-			})
-			return
-		}
-		util.WriteError(w, http.StatusInternalServerError, types.ErrorResponse{
-			ErrCode: "M_UNKNOWN",
-			Message: "Internal server error",
-		})
-		return
-	}
-
-	var valor string
-	if keyName == "displayname" {
-		valor = usuario.Nome
-	} else if keyName == "avatar_url" {
-		valor = usuario.Foto
-	} else {
-		// Se pediu uma chave que não existe no Matrix
-		util.WriteError(w, http.StatusBadRequest, types.ErrorResponse{
-			ErrCode: "M_BAD_JSON",
-			Message: "Invalid profile key",
-		})
-		return
-	}
-
-	response := map[string]string{
-		keyName: valor,
-	}
-
-	util.WriteJSON(w, http.StatusOK, response)
-}
-
-// PUT /_matrix/client/v3/profile/{userId}/{keyName}
-func (h *Handler) putProfileKey(w http.ResponseWriter, r *http.Request) {
-	userID := r.PathValue("userId")
-	keyName := r.PathValue("keyName") // "displayname" ou "avatar_url"
-
-	colunaDB := mapMatrixKeyToDB(keyName)
-	if colunaDB == "" {
-		util.WriteError(w, http.StatusBadRequest, types.ErrorResponse{
-			ErrCode: "M_BAD_JSON",
-			Message: "Invalid profile key",
-		})
-		return
-	}
-
-	var reqBody map[string]string
-	if err := util.ParseBody(r, &reqBody); err != nil {
-		util.WriteError(w, http.StatusBadRequest, types.ErrorResponse{
-			ErrCode: "M_NOT_JSON",
-			Message: "Request body must be JSON",
-		})
-		return
-	}
-
-	novoValor, existe := reqBody[keyName]
-	if !existe {
-		util.WriteError(w, http.StatusBadRequest, types.ErrorResponse{
-			ErrCode: "M_BAD_JSON",
-			Message: "Missing key in request body",
-		})
-		return
-	}
-
-	err := h.userStore.UpdateProfileKey(r.Context(), userID, colunaDB, novoValor)
-	if err != nil {
-		if err == types.ErrNotFound {
-			util.WriteError(w, http.StatusNotFound, types.ErrorResponse{
-				ErrCode: "M_NOT_FOUND",
-				Message: "User not found",
-			})
-			return
-		}
-
-		util.WriteError(w, http.StatusInternalServerError, types.ErrorResponse{
-			ErrCode: "M_UNKNOWN",
-			Message: "Database error",
-		})
-		return
-	}
-
-	util.WriteJSON(w, http.StatusOK, map[string]string{})
-}
-
-// DELETE /_matrix/client/v3/profile/{userId}/{keyName}
-func (h *Handler) deleteProfileKey(w http.ResponseWriter, r *http.Request) {
-	userID := r.PathValue("userId")
-	keyName := r.PathValue("keyName")
-
-	colunaDB := mapMatrixKeyToDB(keyName)
-	if colunaDB == "" {
-		util.WriteError(w, http.StatusBadRequest, types.ErrorResponse{
-			ErrCode: "M_BAD_JSON",
-			Message: "Invalid profile key",
-		})
-		return
-	}
-
-	err := h.userStore.ClearProfileKey(r.Context(), userID, colunaDB)
-
-	if err != nil {
-		util.WriteError(w, http.StatusInternalServerError, types.ErrorResponse{
-			ErrCode: "M_UNKNOWN",
-			Message: "Database error",
-		})
-		return
-	}
-
-	util.WriteJSON(w, http.StatusOK, map[string]string{})
 }
